6-size.c

+++ hello_world/6-size.c
@@ -2,11 +2,11 @@
 
 int main(void)
 {
-	printf ("Size of a char:%d\n byte\(s\)", (int)sizeof(char));
-	printf ("Size of an int:%d\n byte\(s\)", (int)sizeof(int));
-	printf ("Size of a long int:%d\n byte\(s\)",(int)sizeof(long));
-	printf ("Size of a long long int:%d\n byte\(s\)", (int)sizeof(double));
-	printf ("Size of a float:%d\n byte\(s\)", (int)sizeof(float));
+	printf ("Size of a char:%d\n byte\(s\)", sizeof(char));
+	printf ("Size of an int:%d\n byte\(s\)", sizeof(int));
+	printf ("Size of a long int:%d\n byte\(s\)",sizeof(long));
+	printf ("Size of a long long int:%d\n byte\(s\)", sizeof(double));
+	printf ("Size of a float:%d\n byte\(s\)", sizeof(float));
 	
 	return (0);
 }
